+++ lib/main.dart
@@ -21,8 +21,8 @@ import 'package:firebase_ui_auth/firebase_ui_auth.dart';
 
 //Page Imports
 import 'pages/home_page.dart';
-import 'pages/social_media_page.dart';
-import 'pages/historical_data_page.dart';
+import 'pages/leaderboard_page.dart';
+import 'pages/friend_page.dart';
 import 'profile/login_screen.dart';
 
 //Global Variables
